move WeakLruCache to gto-support-androidx-collection library
