+++ gto-support-androidx-collection/src/main/kotlin/org/ccci/gto/android/common/androidx/collection/WeakLruCache.kt
@@ -1,4 +1,4 @@
-package org.ccci.gto.android.common.support.v4.util
+package org.ccci.gto.android.common.androidx.collection
 
 import androidx.collection.ArrayMap
 import androidx.collection.LruCache
