+++ gto-support-core/src/main/kotlin/org/ccci/gto/android/common/support/v4/util/WeakLruCache.kt
@@ -1,0 +1,4 @@
+package org.ccci.gto.android.common.support.v4.util
+
+@Deprecated("Since v3.11.2, This class has been moved to gto-support-androidx-collection")
+typealias WeakLruCache<K, V> = org.ccci.gto.android.common.androidx.collection.WeakLruCache<K, V>
